+++ worker/include/RTC/SvcConsumer.hpp
@@ -25,7 +25,7 @@ namespace RTC
 		void ProducerNewRtpStream(RTC::RtpStream* rtpStream, uint32_t mappedSsrc) override;
 		void ProducerRtpStreamScore(RTC::RtpStream* rtpStream, uint8_t score, uint8_t previousScore) override;
 		void ProducerRtcpSenderReport(RTC::RtpStream* rtpStream, bool first) override;
-		uint16_t GetBitratePriority() const override;
+		uint8_t GetBitratePriority() const override;
 		uint32_t UseAvailableBitrate(uint32_t bitrate, bool considerLoss) override;
 		uint32_t IncreaseLayer(uint32_t bitrate, bool considerLoss) override;
 		void ApplyLayers() override;
